+++ src/media/data/media/contents/movie/internal.py
@@ -7,6 +7,7 @@ from media.xml.namespaces import Namespaces
 from media.data.media.contents.generic.catalog import Catalog
 from media.data.media.contents.genericv.story import Story
 from media.data.media.contents.genericv.crew import Crew
+from media.data.media.contents.movie.classification import Classification
 
 
 class Movie():
@@ -23,6 +24,8 @@ class Movie():
                 self.title = Title(child.text)
             if child.tag == Namespaces.nsf('movie') + 'catalog':
                 self.catalog = Catalog(child)
+            if child.tag == Namespaces.nsf('movie') + 'classification':
+                self.classification = Classification(child)
             if child.tag == Namespaces.nsf('movie') + 'story':
                 self.story = Story(child)
             if child.tag == Namespaces.nsf('movie') + 'description':