+++ CHANGELOG.md
@@ -2,6 +2,7 @@ medialibrary CHANGELOG
 ======================
 
 ## CURRENT
+ - [medialibrary-4](https://github.com/cjcodeproj/medialibrary/issues/4) Basic code for handling movie classifications
  - [medialibrary-2](https://github.com/cjcodeproj/medialibrary/issues/2) Handle description element for non-fictional film
  - Start CHANGELOG.md file
  - Update README.md file